bootstraprag/templates/llamaindex/simple_rag_with_observability/requirements.txt
python-dotenv==1.0.1
llama-index==0.11.7
llama-index-llms-openai==0.2.3
llama-index-llms-ollama==0.3.1
llama-index-embeddings-openai==0.2.4
llama-index-embeddings-ollama==0.3.0
llama-index-vector-stores-qdrant==0.3.0
qdrant-client==1.11.1
arize-phoenix==4.33.1
llama-index-callbacks-arize-phoenix==0.2.1
pydantic==2.9.0
litserve==0.2.2